example/requirements.txt
aiofiles==22.1.0
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
Babel==2.9.1
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
choicelib==0.1.5
click==8.1.3
emoji==2.2.0
frozenlist==1.3.3
idna==3.4
loguru==0.6.0
magic-filter==1.0.9
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pybotterfly @ git+https://github.com/Ninzalo/PyBotterfly.git@c3e7097c815ca6ad552db95a95a7af4a04498fe1
pydantic==1.10.7
python-dotenv==1.0.0
pytz==2023.3
sniffio==1.3.0
typing_extensions==4.5.0
vbml==1.1.post1
vkbottle==4.3.12
vkbottle-types==5.131.146.15
watchfiles==0.18.1
yarl==1.8.2
